+++ app.js
@@ -1,10 +1,12 @@
+require("dotenv").config()
 const express = require("express")
 const redis = require("redis")
 const { APIRouter } = require("./router/api")
 const { successify, errorify } = require("./utils")
+const cors = require("cors")
 
 const app = express()
-
+app.use(cors())
 
 app.use("/api", APIRouter)
 
@@ -12,4 +14,4 @@ app.get("/ping", (req, res) => {
   return res.json(successify({ message: "pong" }))
 })
 
-app.listen(3000, () => console.log("server running at port 3000"))
+app.listen(process.env.PORT, () => console.log(`server running at port ${process.env.PORT}`))
