+++ lib/db.js
@@ -17,10 +17,13 @@ class fastLookup{
   }
   async findPattern(pattern){
     //return await this.search.KEYS(pattern)
-    return await this.search.keys(pattern)
+    return await this.search.keys(`*${pattern}*`) //pattern)
   }
   async map2Meta(id){
     return await this.meta.HGETALL(id)
+  }
+  async map2Id(idxs){
+    return await Promise.all(idxs.map(async (idx) => await this.search.get(idx)))
   }
 }
