--- router/utils/index.js
@@ -1,8 +1,0 @@
-const { randomSample, randInt } = require("./random")
-const { successify, errorify } = require("./response")
-module.exports = {
-  randomSample,
-  randInt,
-  successify,
-  errorify
-}
