--- router/utils/response.js
@@ -1,7 +1,0 @@
-exports.successify = (payload) => {
-  return { success: true, error: false, ...payload}
-}
-
-exports.errorify = (payload) => {
-  return { success: false, ...payload}
-}
